requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
certifi==2023.7.22
charset-normalizer==3.2.0
cytoolz==0.12.2
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
importlib-resources==6.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
protobuf==4.24.0
pycryptodome==3.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.9.2
six==1.16.0
toolz==0.12.0
typing-extensions==4.7.1
urllib3==2.0.4
web3==6.8.0
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
